+++ CHANGELOG.md
@@ -8,8 +8,8 @@ Improvements:
 - Added .NET 7.0 support
 
 Breaking Changes:
-- Upgraded [Castle.Core] version to 5.1.0 (TBD)
-- Upgraded [Castle.Windsor] version to 6.0.0 (TBD)
+- Upgraded [Castle.Core] version to 5.1.1
+- Upgraded [Castle.Windsor] version to 6.0.0
 
 
 ## 5.3.0 (2022-09-17)